+++ src/webapp/src/components/analyze/date-range-picker/DateRangePickerHeader.vue
@@ -1,13 +1,18 @@
 <script>
 import { mapGetters } from 'vuex'
 
+import { EVENTS } from '@/components/analyze/date-range-picker/events'
 import {
   getDateLabel,
   getHasValidDateRange
 } from '@/components/analyze/date-range-picker/utils'
+import DateRangeCustomVsRelative from '@/components/analyze/date-range-picker/DateRangeCustomVsRelative'
 
 export default {
   name: 'DateRangePickerHeader',
+  components: {
+    DateRangeCustomVsRelative
+  },
   props: {
     attributePair: { type: Object, required: true },
     attributePairsModel: { type: Array, required: true }
@@ -49,21 +54,21 @@ export default {
           attributePair.attribute.sourceName === sourceName &&
           attributePair.attribute.name === name
       )
-      this.$emit('attribute-pair-change', targetAttributePair)
+      this.$emit(EVENTS.ATTRIBUTE_PAIR_CHANGE, targetAttributePair)
     },
     onClearDateRange() {
-      this.$emit('clear-date-range', this.attributePair)
+      this.$emit(EVENTS.CLEAR_DATE_RANGE, this.attributePair)
     }
   }
 }
 </script>
 
 <template>
-  <div class="columns is-vcentered">
-    <div class="column">
+  <div>
+    <div class="mb1r">
       <div class="field">
         <div class="control">
-          <span class="select is-small">
+          <span class="select is-fullwidth">
             <select
               :value="getValue(attributePair)"
               :class="{
@@ -84,22 +89,28 @@ export default {
       </div>
     </div>
 
-    <div class="column">
-      <div class="is-flex is-vcentered is-pulled-right">
-        <span
-          class="is-size-7"
-          :class="{
-            'mr-05r': getHasValidDateRange(attributePair.dateRange)
-          }"
-          >{{ getDateLabel(attributePair) }}</span
-        >
-        <button
-          v-if="getHasValidDateRange(attributePair.dateRange)"
-          class="button is-small"
-          @click="onClearDateRange"
-        >
-          Clear
-        </button>
+    <div class="columns is-vcentered">
+      <div class="column">
+        <DateRangeCustomVsRelative :attribute-pair="attributePair" />
+      </div>
+
+      <div class="column">
+        <div class="is-flex is-vcentered is-pulled-right">
+          <span
+            class="is-size-7"
+            :class="{
+              'mr-05r': getHasValidDateRange(attributePair.absoluteDateRange)
+            }"
+            >{{ getDateLabel(attributePair) }}</span
+          >
+          <button
+            v-if="getHasValidDateRange(attributePair.absoluteDateRange)"
+            class="button is-small"
+            @click="onClearDateRange"
+          >
+            Clear
+          </button>
+        </div>
       </div>
     </div>
   </div>