+++ src/webapp/src/components/analyze/date-range-picker/events.js
@@ -1,0 +1,5 @@
+export const EVENTS = Object.freeze({
+  ATTRIBUTE_PAIR_CHANGE: 'attribute-pair-change',
+  CLEAR_DATE_RANGE: 'clear-date-range',
+  CHANGE_DATE_RANGE: 'change-date-range'
+})